requirements.txt
azure-iot-device~=2.12.0
Adafruit_DHT==1.3.3